requirements.txt
flask
flask_sqlalchemy
flask_wtf
flask_bcrypt
flask_login
flask_migrate
pymysql
python-dotenv
